修复 获取社团问题

+++ src/main/java/top/pdev/you/interfaces/model/vo/AssociationInfoVO.java
@@ -17,5 +17,5 @@ public class AssociationInfoVO {
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Integer status;
-    private Integer numbers;
+    private Long numbers;
 }
